# docs/swagger.yaml
basePath: /api
definitions:
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.JobSubmitRequest:
    properties:
      callback_url:
        type: string
      content:
        type: string
      credential_id:
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      pipeline_id:
        type: integer
      skip_cache:
        type: boolean
      source_url:
        type: string
      type:
        type: string
    required:
    - pipeline_id
    - type
    type: object
  service.LLMProviderCreateRequest:
    properties:
      active:
        type: boolean
      api_key:
        type: string
      base_url:
        type: string
      default_model:
        type: string
      display_name:
        type: string
      input_price:
        type: number
      is_default:
        type: boolean
      max_concurrency:
        type: integer
      name:
        type: string
      output_price:
        type: number
    required:
    - api_key
    - name
    type: object
  service.LLMProviderUpdateRequest:
    properties:
      active:
        type: boolean
      api_key:
        type: string
      base_url:
        type: string
      default_model:
        type: string
      display_name:
        type: string
      input_price:
        type: number
      max_concurrency:
        type: integer
      output_price:
        type: number
    type: object
...
host: localhost:7847
info:
  contact: {}
  description: LLM 文档处理与知识图谱 API 中间件
  title: Nexus API
  version: "1.0"
paths:
...
  /admin/jobs:
    get:
      description: 分页获取任务列表，支持按状态筛选
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取任务列表
      tags:
      - 任务
  /admin/jobs/{id}/retry:
    post:
      description: 重试失败的任务
      parameters:
      - description: 任务 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 重试任务
      tags:
      - 任务
  /admin/jobs/recommend:
    get:
      description: 根据场景推荐相关任务
      parameters:
      - description: 场景名称
        in: query
        name: scene
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 推荐任务
      tags:
      - 任务
  /admin/llm-providers:
    get:
      description: 获取所有已配置的 LLM 提供商
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 LLM Provider 列表
      tags:
      - LLM Provider
    post:
      consumes:
      - application/json
      description: 添加新的 LLM 提供商配置
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LLMProviderCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建 LLM Provider
      tags:
      - LLM Provider
  /admin/llm-providers/{id}:
    delete:
      description: 删除指定 LLM 提供商
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除 LLM Provider
      tags:
      - LLM Provider
    put:
      consumes:
      - application/json
      description: 更新 LLM 提供商配置
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LLMProviderUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新 LLM Provider
      tags:
      - LLM Provider
  /admin/llm-providers/{id}/default:
    put:
      description: 将指定 Provider 设为默认
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 设置默认 Provider
      tags:
      - LLM Provider
  /admin/llm-providers/{name}/models:
    get:
      description: 获取指定 Provider 支持的模型
      parameters:
      - description: Provider 名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 Provider 模型列表
      tags:
      - LLM Provider
...
  /v1/jobs:
    post:
      consumes:
      - application/json
      description: 提交文档到 Pipeline 进行异步处理，返回任务 ID
      parameters:
      - description: 任务提交请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.JobSubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 提交处理任务
      tags:
      - 任务
  /v1/jobs/{id}:
    get:
      description: 根据任务 UUID 查询处理状态和结果
      parameters:
      - description: 任务 UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 查询任务状态
      tags:
      - 任务
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"